+++ R/polar_annulusMap.R
@@ -7,6 +7,7 @@
 #' interacted with. Using the `static` argument allows for static images to be
 #' produced instead.
 #'
+#' @inheritSection polarMap Parallel processing with mirai
 #' @inheritSection polarMap Customisation of static maps using ggplot2
 #' @family directional analysis maps
 #'
@@ -63,7 +64,6 @@ annulusMap <- function(
   static = FALSE,
   static.nrow = NULL,
   progress = TRUE,
-  n.core = 1L,
   ...,
   control = NULL
 ) {
@@ -204,8 +204,7 @@ annulusMap <- function(
       d.fig = d.fig,
       popup = popup,
       label = label,
-      progress = progress,
-      ncores = n.core
+      progress = progress
     )
 
   if (static) {