+++ R/polar_pollroseMap.R
@@ -6,6 +6,7 @@
 #' maps are dynamic and can be panned, zoomed, and otherwise interacted with.
 #' Using the `static` argument allows for static images to be produced instead.
 #'
+#' @inheritSection polarMap Parallel processing with mirai
 #' @inheritSection polarMap Customisation of static maps using ggplot2
 #' @family directional analysis maps
 #'
@@ -75,7 +76,6 @@ pollroseMap <- function(
   static = FALSE,
   static.nrow = NULL,
   progress = TRUE,
-  n.core = 1L,
   ...,
   control = NULL
 ) {
@@ -176,8 +176,7 @@ pollroseMap <- function(
       d.fig = d.fig,
       popup = popup,
       label = label,
-      progress = progress,
-      ncores = n.core
+      progress = progress
     )
 
   if (static) {